backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
pymupdf==1.23.8
python-docx==1.1.0
docx2txt==0.8
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.0
groq==0.2.0
huggingface-hub==0.17.3
transformers==4.35.0
torch==2.1.0
supabase==2.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.24.1
aiofiles==23.2.1
